Updated sam_bam_verify.c  for the following functionality:
-- Formerly, would produce output for every unmapped read. Will now only produce output for the first unmapped read in a file.
(note : it appears that samtools will still produce SOME output for every read mapped to a 'broken' chromosome, but sam_bam_verify won't.

-- Added a metric to the output, "Unique multiply-mapped reads"
This metric represents the number of READ IDs that are mapped to multiple locationes on the chromosome.
It increments by one for each  unique read id + pair #, no matter how many times they are mapped.

SO, each read ID can contribute a max of 2 to this metric, one for each member of the pair (should it be paired.)

This can work with the "Unique mapped reads" (which counts both uniquely-mapped and multiply-mapped reads) to get the
uniquely-mapped reads:

uniquely mapped reads = unique mapped reads - unique multiply-mapped reads.


FUNCTIONALITY CHANGED:

Previously, if there was a read that was unmapped (per the flag), but DID have a valid chromosome (not * ) listed,
it would be included in the output file and would contribute to the "Unique total reads" metric.

We consider this to be incorrect, and those are now dropped from the file.